requirements.txt
aiocache==0.3.2
aioconsole==0.1.2
aiohttp==2.0.5
aiomcache==0.5.1
aioredis==0.3.0
appdirs==1.4.3
async-timeout==1.2.0
asyncpg==0.9.0
cffi==1.10.0
chardet==2.3.0
guillotina==1.0.0a3
hiredis==0.2.0
jsonschema==2.6.0
multidict==2.1.4
packaging==16.8
pycparser==2.17
pycrypto==2.6.1
PyJWT==1.4.2
pyparsing==2.2.0
python-dateutil==2.6.0
requests==2.13.0
six==1.10.0
ujson==1.35
yarl==0.10.0
zope.interface==4.3.3
